feat: New CLI option to change badge style. Refactor readmeai package structure to improve maintainability.
