+++ readmeai/cli/options.py
@@ -6,13 +6,24 @@ import os
 
 import click
 
+BADGE_CHOICES = ("shields", "square")
+
+
 api_key = click.option(
     "-k",
     "--api-key",
     default=os.environ.get("OPENAI_API_KEY", None),
     help="Large language model API secret key.",
 )
-
+badges = click.option(
+    "-b",
+    "--badges",
+    type=click.Choice(BADGE_CHOICES, case_sensitive=False),
+    default="shields",
+    help="""Badge icon type to use in README.md header. \
+        - 'shields' refers to badges from shields.io \
+        - 'square' refers to app-like square badges.""",
+)
 encoding = click.option(
     "-c",
     "--encoding",
