+++ readmeai/data/skill_icons.json
@@ -1,0 +1,204 @@
+{
+    "icons": {
+        "names": [
+            "ableton",
+            "activitypub",
+            "actix",
+            "adonis",
+            "ae",
+            "aiscript",
+            "alpinejs",
+            "androidstudio",
+            "angular",
+            "ansible",
+            "apollo",
+            "appwrite",
+            "arduino",
+            "astro",
+            "atom",
+            "au",
+            "autocad",
+            "aws",
+            "azul",
+            "azure",
+            "babel",
+            "bash",
+            "bevy",
+            "blender",
+            "bootstrap",
+            "bsd",
+            "c",
+            "cs",
+            "cpp",
+            "crystal",
+            "cassandra",
+            "clojure",
+            "cloudflare",
+            "cmake",
+            "codepen",
+            "coffeescript",
+            "css",
+            "d3",
+            "dart",
+            "deno",
+            "devto",
+            "discord",
+            "bots",
+            "django",
+            "docker",
+            "dotnet",
+            "dynamodb",
+            "eclipse",
+            "electron",
+            "elixir",
+            "emacs",
+            "ember",
+            "emotion",
+            "express",
+            "fastapi",
+            "fediverse",
+            "figma",
+            "firebase",
+            "flask",
+            "flutter",
+            "forth",
+            "fortran",
+            "gamemakerstudio",
+            "gatsby",
+            "gcp",
+            "git",
+            "github",
+            "githubactions",
+            "gitlab",
+            "gherkin",
+            "go",
+            "gradle",
+            "godot",
+            "grafana",
+            "graphql",
+            "gtk",
+            "gulp",
+            "haskell",
+            "haxe",
+            "haxeflixel",
+            "heroku",
+            "hibernate",
+            "html",
+            "idea",
+            "ai",
+            "instagram",
+            "ipfs",
+            "java",
+            "js",
+            "jenkins",
+            "jest",
+            "jquery",
+            "kafka",
+            "kotlin",
+            "ktor",
+            "kubernetes",
+            "laravel",
+            "latex",
+            "linkedin",
+            "linux",
+            "lit",
+            "lua",
+            "md",
+            "mastodon",
+            "materialui",
+            "matlab",
+            "maven",
+            "misskey",
+            "mongodb",
+            "mysql",
+            "neovim",
+            "nestjs",
+            "netlify",
+            "nextjs",
+            "nginx",
+            "nim",
+            "nodejs",
+            "nuxtjs",
+            "ocaml",
+            "octave",
+            "openshift",
+            "openstack",
+            "perl",
+            "ps",
+            "php",
+            "plan9",
+            "planetscale",
+            "postgres",
+            "postman",
+            "powershell",
+            "pr",
+            "prisma",
+            "processing",
+            "prometheus",
+            "pug",
+            "py",
+            "pytorch",
+            "qt",
+            "r",
+            "rabbitmq",
+            "rails",
+            "raspberrypi",
+            "react",
+            "reactivex",
+            "redis",
+            "redux",
+            "regex",
+            "remix",
+            "replit",
+            "rocket",
+            "rollupjs",
+            "ros",
+            "ruby",
+            "rust",
+            "sass",
+            "spring",
+            "sqlite",
+            "stackoverflow",
+            "styledcomponents",
+            "supabase",
+            "scala",
+            "selenium",
+            "sentry",
+            "sequelize",
+            "sketchup",
+            "solidity",
+            "solidjs",
+            "svelte",
+            "svg",
+            "swift",
+            "symfony",
+            "tailwind",
+            "tauri",
+            "tensorflow",
+            "threejs",
+            "twitter",
+            "ts",
+            "unity",
+            "unreal",
+            "v",
+            "vala",
+            "vercel",
+            "vim",
+            "visualstudio",
+            "vite",
+            "vscode",
+            "vue",
+            "wasm",
+            "webflow",
+            "webpack",
+            "windicss",
+            "wordpress",
+            "workers",
+            "xd",
+            "zig"
+        ]
+    },
+    "url": {
+        "base_url": "https://skillicons.dev/icons?i="
+    }
+}
