+++ readmeai/settings/config.toml
@@ -16,12 +16,13 @@ name = "readme-ai"
 
 # Paths
 [paths]
-badges = "conf/badges.json"
+output = "readme-ai.md"
+badge_icons = "data/shields_icons.json"
+skill_icons = "data/skill_icons.json"
 dependency_files = "dependency_files.toml"
 ignore_files = "ignore_files.toml"
 language_names = "language_names.toml"
 language_setup = "language_setup.toml"
-readme = "readme-ai.md"
 
 # Prompts
 [prompts]
@@ -64,20 +65,34 @@ dropdown = """<details closed><summary>{}</summary>\n\n{}\n\n</details>\n"""
 header = """<div align="center">
 <h1 align="center">
 <img src="https://raw.githubusercontent.com/PKief/vscode-material-icon-theme/ec559a9f6bfd399b82bb44393651661b08aaf7ba/icons/folder-markdown-open.svg" width="100" />
-<br>{}
-</h1>
-<h3>◦ {}</h3>
+<br>{0}</h1>
+<h3>◦ {1}</h3>
 <h3>◦ Developed with the software and tools below.</h3>
 """
 badges = """<p align="center">
 {0}
 </p>
-<img src="https://img.shields.io/github/languages/top/{1}?style&color=5D6D7E" alt="GitHub top language" />
-<img src="https://img.shields.io/github/languages/code-size/{1}?style&color=5D6D7E" alt="GitHub code size in bytes" />
-<img src="https://img.shields.io/github/commit-activity/m/{1}?style&color=5D6D7E" alt="GitHub commit activity" />
 <img src="https://img.shields.io/github/license/{1}?style&color=5D6D7E" alt="GitHub license" />
+<img src="https://img.shields.io/github/last-commit/{1}?style&color=5D6D7E" alt="git-last-commit" />
+<img src="https://img.shields.io/github/commit-activity/m/{1}?style&color=5D6D7E" alt="GitHub commit activity" />
+<img src="https://img.shields.io/github/languages/top/{1}?style&color=5D6D7E" alt="GitHub top language" />
 </div>
 \n---
+"""
+badges_alt = """
+<div align="center">
+    <h1>{}</h1>
+    <h3>◦ {}</h3>
+    <h3>◦ Developed with the software and tools below</h3>
+</div>
+
+<p align="center">
+  <a href="https://skillicons.dev">
+    <img src={} />
+  </a>
+</p>
+
+---
 """
 toc = """
 ## 📖 Table of Contents
@@ -179,5 +194,16 @@ This project is licensed under the `ℹ️  LICENSE-TYPE` License. See the [LICE
 
 `- ℹ️ List any resources, contributors, inspiration, etc.`
 
+[↑ Return](#Top)
+
 ---
+"""
+# New Markdown Template Code
+header_png = """[<img src="https://img.icons8.com/?size=512&id=55494&format=png" height="120" align="left">](https://img.icons8.com/?size=512&id=55494&format=png)
+
+# {}
+
+<br>
+
+## {}
 """
