+++ readmeai/utils/utils.py
@@ -1,0 +1,63 @@
+"""Utility methods for the readme-ai application."""
+
+import re
+from typing import List
+
+from readmeai.core import logger
+
+logger = logger.Logger(__name__)
+
+
+def is_valid_url(url: str) -> bool:
+    """Check if a given string is a valid URL."""
+    regex = re.compile(
+        r"^(?:http|ftp)s?://"
+        r"(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,63}|[A-Z]{2,63}\.[A-Z]{2,63}))"
+        r"(?::\d+)?"
+        r"(?:/?|[/?]\S+)$",
+        re.IGNORECASE,
+    )
+    return bool(regex.match(url))
+
+
+def flatten_list(nested_list: List) -> List:
+    """Flattens a nested list."""
+    result = []
+    for item in nested_list:
+        if isinstance(item, list):
+            result.extend(flatten_list(item))
+        else:
+            result.append(item)
+    return result
+
+
+def format_sentence(text: str) -> str:
+    """Clean and format the generated text from the model."""
+    # Remove newlines and tabs
+    text = text.replace("\n", "").replace("\t", "")
+
+    # Remove non-letter characters from the beginning of the string
+    text = re.sub(r"^[^a-zA-Z]*", "", text)
+
+    # Remove extra white space around punctuation except for '('
+    text = re.sub(r"\s*([)'.!,?;:])(?!\.\s*\w)", r"\1", text)
+
+    # Remove extra white space before opening parentheses
+    text = re.sub(r"(\()\s*", r"\1", text)
+
+    # Replace multiple consecutive spaces with a single space
+    text = re.sub(r" +", " ", text)
+
+    # Remove extra white space around hyphens
+    text = re.sub(r"\s*-\s*", "-", text)
+
+    return text.strip().strip('"')
+
+
+def remove_substring(input_string: str) -> str:
+    """Remove text between HTML tags."""
+    pattern = r"</p>.*?</div>"
+    output_string = re.sub(
+        pattern, "</p>\n</div>", input_string, flags=re.DOTALL
+    )
+    return output_string
